Adding first cut at annotated assembly for fib.c. (Work in progress)
The fib-clang.s file was generated by clang (with -g -S flags). I will
annotate this to explain what it does. While any hand-written assembly
code will look quite different from compiler-generated assembler, it
seems instructive to start from compiler-generated code as an exercise in
learning. It is also useful to know how to step through compiler-generated
code if you are going to be doing low-level debugging of compiled code.

The fib-gdb-dump.s code was generated from the "disassemble /m" command
in gdb. Note the differences in what the gdb disassembler shows as
compared to the CLANG assembler source. Although the opcodes are likely
to be the same (I haven't checked yet), the way they are displayed is
somewhat different, especially when referring to labels (shown as an
offset from the PC rather than as a label, since labels are often lost
in compilation).